+++ changelog.md
@@ -1,3 +1,5 @@
+- check for error/empty response when checking current branch
+
 # 0.3.0
 - add support for fetching services as strings [#3]
 - support setting a base_url for select services. this adds support for github enterprise (and in the future gitlab)